DB refactor

+++ app/src/main/java/com/onbiron/forecastmvvm/data/network/response/future/Current.kt
@@ -7,7 +7,7 @@ data class Current(
     val clouds: Double,
     @SerializedName("dew_point")
     val dewPoint: Double,
-    val dt: Double,
+    val dt: Long,
     @SerializedName("feels_like")
     val feelsLike: Double,
     val humidity: Double,
