+++ app/src/main/java/com/onbiron/forecastmvvm/data/network/response/future/Minutely.kt
@@ -1,9 +1,6 @@
 package com.onbiron.forecastmvvm.data.network.response.future
 
-
-import com.google.gson.annotations.SerializedName
-
 data class Minutely(
-    val dt: Double,
+    val dt: Long,
     val precipitation: Double
 )
